refactor: modularization of extract engine
