+++ src/h5radiomics/utils/__init__.py
@@ -2,5 +2,9 @@ from .config import *
 from .h5 import *
 from .io import *
 from .paths import *
+from .patchio import * 
+from .feature_utils import * 
+from .maskgeo import * 
 
 __all__ = []
+
